generate_*_module: don't build for experimental branches
Don't build from experimental branches. Otherwise we'd need to have some
mechanism to publish experimental module builds, which we don't have for
Java and JavaScript modules.

Maybe in the future we might have something like experimental Maven or
NPM repositories, then we'd want to build experimental branches for
them. But for now, just skip and avoid polluting the "master" builds
repositories.

+++ kurento_generate_java_module.sh
@@ -33,6 +33,18 @@ kurento_check_version.sh false || {
   exit 1
 }
 
+# Don't build from experimental branches. Otherwise we'd need to have some
+# mechanism to publish experimental module builds, which we don't have for
+# Java and JavaScript modules.
+#
+# Maybe in the future we might have something like experimental Maven or NPM
+# repositories, then we'd want to build experimental branches for them. But
+# for now, just skip and avoid polluting the "master" builds repositories.
+if [[ "$JOB_GIT_NAME" != "master" ]]; then
+  log "Skip building from experimental branch '$JOB_GIT_NAME'"
+  exit 0
+}
+
 rm -rf build
 mkdir build && cd build
 cmake .. -DGENERATE_JAVA_CLIENT_PROJECT=TRUE -DDISABLE_LIBRARIES_GENERATION=TRUE || {
